Fixed bug when day isn't valid in resulting month
For some months, 31 isn't a valid day (Feb, Apr, Jun, Sep, Nov), and for
February, 30 or 31 isn't valid, and 29 is only valid in a leap year.

Now check the resulting year and month to set the day to either 30 (for
Apr, Jun, Sep, nov) or 29 (Feb in leap year) or 28 (Feb in normal year).